+++ packages/commands/src/start.ts
@@ -9,7 +9,7 @@ import { initLocalnet } from "../../localnet/src";
 import { isDockerAvailable } from "../../localnet/src/isDockerAvailable";
 import { isSolanaAvailable } from "../../localnet/src/isSolanaAvailable";
 import { isSuiAvailable } from "../../localnet/src/isSuiAvailable";
-import { tonStart } from "../../localnet/src/tonStart";
+import { start as tonStart } from "../../localnet/src/ton/setup";
 import { initLocalnetAddressesSchema } from "../../types/zodSchemas";
 
 const LOCALNET_JSON_FILE = "./localnet.json";
@@ -185,7 +185,7 @@ const startLocalnet = async (options: {
     process.exit(0);
   }
 
-  await new Promise(() => {});
+  await new Promise(() => { });
 };
 
 export const startCommand = new Command("start")
