+++ packages/tasks/src/localnet.ts
@@ -9,7 +9,7 @@ import { initLocalnet } from "../../localnet/src";
 import { isDockerAvailable } from "../../localnet/src/isDockerAvailable";
 import { isSolanaAvailable } from "../../localnet/src/isSolanaAvailable";
 import { isSuiAvailable } from "../../localnet/src/isSuiAvailable";
-import { tonStart } from "../../localnet/src/tonStart";
+import { start as tonStart } from "../../localnet/src/ton/setup";
 
 const LOCALNET_JSON_FILE = "./localnet.json";
